# docs/swagger.yaml
basePath: /
definitions:
...
  models.AdminDetail:
    properties:
      email:
        type: string
      password:
        maxLength: 20
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
...
  models.SignupDetail:
    properties:
      email:
        type: string
      firstname:
        type: string
      lastname:
        type: string
      password:
        type: string
      phone:
        type: string
      referral_code:
        type: string
    required:
    - email
    - firstname
    - lastname
    - password
    - phone
    type: object
...
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
host: www.zogfestiv.store
info:
  contact: {}
  description: API for ecommerce website
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Zog_festiv eCommerce API
  version: "1.0"
paths:
  /admin/admin-login:
    post:
      consumes:
      - application/json
      description: Login handler for admin
      parameters:
      - description: Admin login details
        in: body
        name: admin
        required: true
        schema:
          $ref: '#/definitions/models.AdminDetail'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Admin Login
      tags:
      - Admin Authentication
...
  /admin/dashboard:
    get:
      consumes:
      - application/json
      description: Get Amin Home Page with Complete Details
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Admin Dashboard
      tags:
      - Admin Dash Board
...
  /admin/orders/approve-order/{id}:
    get:
      consumes:
      - application/json
      description: Approve Order from admin side which is in processing state
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Approve Order
      tags:
      - Admin Order Management
  /admin/orders/cancel-order/{id}:
    get:
      consumes:
      - application/json
      description: Cancel Order from admin side
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Cancel Order Admin
      tags:
      - Admin Order Management
...
  /admin/users:
    post:
      consumes:
      - application/json
      description: Create a new user from admin side
      parameters:
      - description: Add a new user
        in: body
        name: userDetails
        required: true
        schema:
          $ref: '#/definitions/models.SignupDetail'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Create User By Admin
      tags:
      - Admin User Management
  /admin/users/{page}:
    get:
      consumes:
      - application/json
      description: Retrieve users with pagination to admin side
      parameters:
      - description: Page number
        in: path
        name: page
        required: true
        type: string
      - description: page size
        in: query
        name: pageSize
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Get Users Details To Admin
      tags:
      - Admin User Management
  /admin/users/block-users/{id}:
    post:
      consumes:
      - application/json
      description: Block an existing user using user id
      parameters:
      - description: user-id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Block  user
      tags:
      - Admin User Management
  /admin/users/unblock-users/{id}:
    post:
      consumes:
      - application/json
      description: Unblock an already blocked user using user id
      parameters:
      - description: user-id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Unblock  User
      tags:
      - Admin User Management
...
schemes:
- http
securityDefinitions:
  JWT:
    in: header
    name: token
    type: apiKey
swagger: "2.0"